source/requirements.txt
mwparserfromhell==0.6.6
pymongo==4.10.1
logging
python-dotenv
